workdir/requirements.txt
Django~=5.2.0
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.5.0
django-guardian~=3.0.0
psycopg2-binary>=2.9,<3.0
redis>=4.0,<5.0
django-redis>=5.0,<6.0
python-dotenv>=0.20,<1.1
dj-database-url>=0.5,<1.0
gunicorn
drf-spectacular>=0.26,<0.28
drf-spectacular-sidecar>=2024.7.1
sentry-sdk[django,celery]>=1.0,<2.0
django-storages[boto3]>=1.13,<1.15
django-filter>=23.0,<25.0
django-allauth~=0.58.2
dj-rest-auth~=5.0.0
beautifulsoup4>=4.12,<4.13
lxml>=4.9,<5.2
